requirements.txt
aniso8601==4.0.1
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-Testing==0.7.1
funcsigs==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==4.3.0
nose==1.3.7
pathlib2==2.3.2
pluggy==0.8.0
psycopg2-binary==2.7.6.1
py==1.7.0
pytest==4.0.1
pytz==2018.7
scandir==1.9.0
six==1.11.0
Werkzeug==0.14.1
